+++ Hearthstone Deck Tracker/Hearthstone/RelatedCardsSystem/Cards/Rogue/GaronaHalforcenTheKingslayers.cs
@@ -8,5 +8,6 @@ public class GaronaHalforcenTheKingslayers : ICardWithHighlight
 	public string GetCardId() => HearthDb.CardIds.NonCollectible.Rogue.GaronaHalforcen_TheKingslayersToken;
 
 	public HighlightColor ShouldHighlight(Card card, IEnumerable<Card> deck) =>
-		HighlightColorHelper.GetHighlightColor(card.Rarity == Rarity.LEGENDARY || card.GetTag(GameTag.ELITE) > 0);
+		HighlightColorHelper.GetHighlightColor(card.Rarity == Rarity.LEGENDARY ||
+		                                       (card.Rarity == Rarity.INVALID && card.GetTag(GameTag.ELITE) > 0));
 }