+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fixed The Kingslayer not Highlighting Incindius' Eruptions.
 - Added cards offered by Waveshaping to the bottom of the deck.
 - Added cards discovered by Kaldorei Cultivator to the bottom of the deck.
+- Fixed crashes when Murozond, Unbounded's attack was set to infinity. HDT is now capable of handling all its power.
 
 
 ## **Release v1.48.2 - 2025-10-31**